Delete my attempt to edit the homepage
I tried to edit the "home" layout by changing it in the gem file and then pushing the gem file. The problem is that the gem file (which was in vendor), is normally ignored when I push. It also didn't work - the way the site is generated, me adding an edited gem file did not make the site change.